src/shared/pyproject.toml
[tool.poetry]
name = "shared"
version = "0.1.0"
description = ""
authors = ["Kazox6 <pawelka00florian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
coc-py = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
